BST-57037 Lease or agreement CYA for pages `about-your-landlord`, `current-annual-rent`. Trait `PrintableAddress` with methods to display address
