+++ conf/messages
@@ -851,7 +851,10 @@ checkYourAnswersAboutFranchiseOrLettings.lettingTypeOfBusiness = Description of 
 checkYourAnswersAboutFranchiseOrLettings.annualRent = Annual rent payment
 checkYourAnswersAboutFranchiseOrLettings.dateSumFixed = When sum was fixed?
 checkYourAnswersAboutFranchiseOrLettings.itemsInRent = Items included in rent
-checkYourAnswersAboutYourLeaseOrTenure.heading = This is a holding page for the about your lease or tenure check your answers page
+checkYourAnswersAboutYourLeaseOrTenure.heading = Check your answers - your lease or agreement
+checkYourAnswersAboutYourLeaseOrTenure.yourLandlord.heading = About your landlord
+checkYourAnswersAboutYourLeaseOrTenure.yourRent.heading = Your rent
+checkYourAnswersAboutYourLeaseOrTenure.currentAnnualRent = Current annual rent
 checkYourAnswersAboutTheProperty.heading = Check your answers - about you and the property
 checkYourAnswersAboutTheProperty.aboutYou.heading = Your contact details
 checkYourAnswersAboutTheProperty.aboutProperty.heading = How the property is used
